+++ src/partitioner/geo-data.cpp
@@ -8,10 +8,10 @@ partitioner::GeoData::GeoData()
 
 partitioner::GeoData::GeoData(std::string latitude, std::string longitude)
 {
-    this->latitude = load_vector<int>(latitude);
-    this->longitude = load_vector<int>(longitude);
+    this->latitude = load_vector<float>(latitude);
+    this->longitude = load_vector<float>(longitude);
 }
 
-partitioner::GeoData::GeoData(std::vector<int> lat, std::vector<int> lon) : latitude(lat), longitude(lon)
+partitioner::GeoData::GeoData(std::vector<float> lat, std::vector<float> lon) : latitude(lat), longitude(lon)
 {
 }
